Sign in now  checks user after authentication
